Rudimentary templating support
It is now possible to instantiate new zones from existing templates.
This is about all my automatons need to do with templates.

It is not possible to create new templates, or manipulate existing
templates, in any way using this library.  grep for NotImplementedError
to locate the stubs that need to be filled out if you want this
functionality.  Don't forget to remove the comment at the top of
tests/test_template.py.