Do not import version
While this practice is convenient and all, importing the module this way
means that my packaging environment needs to have more dependencies
installed on it than should be necessary.  To import __version__ from
__init__.py, Python tries to parse and compile my modules, which in turn
pulls in lxml and friends.  I am not specifying lxml as a build
dependency just so I can avoid having to specify a version in two
places.

There is probably a nicer way of going about all this.